backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-jazzmin>=3.0,<3.1
psycopg[binary]>=3.2,<3.3
celery>=5.4,<5.5
redis>=5.1,<5.2
gunicorn>=23.0,<24.0
whitenoise>=6.8,<6.9
reportlab>=4.2,<4.3
